+++ meta_package_manager/pool.py
@@ -50,7 +50,7 @@ from .managers.pacman import Pacaur, Pacman, Paru, Yay
 from .managers.pacstall import Pacstall
 from .managers.pip import Pip
 from .managers.pipx import Pipx
-from .managers.pkg import PKG
+from .managers.pkg import PKG, Ports
 from .managers.scoop import Scoop
 from .managers.sdkman import SDKMAN
 from .managers.sfsu import SFSU
@@ -105,6 +105,7 @@ manager_classes = (
     Pip,
     Pipx,
     PKG,
+    Ports,
     Scoop,
     SDKMAN,
     SFSU,